services/identity-service/requirements.txt
django==5.0.1
strawberry-graphql[django]
psycopg2-binary
requests